+++ BambooTracker/playback.cpp
@@ -124,7 +124,8 @@ void PlaybackManager::startPlay()
 	tickCounter_.lock()->setTempo(song.getTempo());
 	tickCounter_.lock()->setSpeed(song.getSpeed());
 	tickCounter_.lock()->setGroove(mod_.lock()->getGroove(song.getGroove()).getSequence());
-	tickCounter_.lock()->setGrooveEnebled(!song.isUsedTempo());
+	tickCounter_.lock()->setGrooveTrigger(song.isUsedTempo() ? GrooveTrigger::Invalid
+															 : GrooveTrigger::ValidByGlobal);
 	tickCounter_.lock()->resetCount();
 	tickCounter_.lock()->setPlayState(true);
 
@@ -1091,19 +1092,19 @@ bool PlaybackManager::effPatternBreak(int nextStep)
 void PlaybackManager::effSpeedChange(int speed)
 {
 	tickCounter_.lock()->setSpeed(speed ? speed : 1);
-	tickCounter_.lock()->setGrooveEnebled(false);
+	tickCounter_.lock()->setGrooveTrigger(GrooveTrigger::Invalid);
 }
 
 void PlaybackManager::effTempoChange(int tempo)
 {
 	tickCounter_.lock()->setTempo(tempo);
-	tickCounter_.lock()->setGrooveEnebled(false);
+	tickCounter_.lock()->setGrooveTrigger(GrooveTrigger::Invalid);
 }
 
 void PlaybackManager::effGrooveChange(int num)
 {
 	tickCounter_.lock()->setGroove(mod_.lock()->getGroove(num).getSequence());
-	tickCounter_.lock()->setGrooveEnebled(true);
+	tickCounter_.lock()->setGrooveTrigger(GrooveTrigger::ValidByLocal);
 }
 
 void PlaybackManager::readTick(int rest)